requirements.txt
appscript==1.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docx2pdf==0.1.8
Flask==3.1.0
Flask-WTF==1.2.2
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
pyperclip==1.9.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
